PR fixups
